refactor: ShowingController와 ManipulatingController를 Controller로 통합
