+++ app/src/main/java/com/codesoom/assignment/controllers/RequestParamValidation.java
@@ -1,11 +1,8 @@
 package com.codesoom.assignment.controllers;
 
-import com.codesoom.assignment.controllers.dtos.TaskRequestDto;
 import com.codesoom.assignment.interfaces.DefaultRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
-
-import java.util.Objects;
 
 public class RequestParamValidation {
     private final Long pathVariable;
