+++ app/src/main/java/com/codesoom/assignment/interfaces/DefaultController.java
@@ -1,5 +1,6 @@
 package com.codesoom.assignment.interfaces;
 
+import com.codesoom.assignment.controllers.dtos.TaskRequestDto;
 import com.codesoom.assignment.controllers.dtos.TaskResponseDto;
 
 import java.util.List;
@@ -9,5 +10,9 @@ public interface DefaultController {
 
     TaskResponseDto showBy(Long id);
 
-    ManipulatingController manipulator();
+    TaskResponseDto create(TaskRequestDto requestDto);
+
+    TaskResponseDto update(Long id, TaskRequestDto requestDto);
+
+    void deleteBy(Long id);
 }
